requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-phonenumber-field==7.1.0
idna==3.4
kavenegar==1.1.2
psycopg2-binary==2.9.6
pyotp==2.8.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
